+++ src/promptflow/tests/sdk_cli_test/unittests/test_utils.py
@@ -4,6 +4,7 @@
 
 import argparse
 import importlib
+import json
 import os
 import shutil
 import sys
@@ -16,6 +17,7 @@ import datetime
 import mock
 import pandas as pd
 import pytest
+import time
 
 from promptflow._cli._params import AppendToDictAction
 from promptflow._cli._utils import (
@@ -23,7 +25,7 @@ from promptflow._cli._utils import (
     _calculate_column_widths,
     list_of_dict_to_nested_dict,
 )
-from promptflow._constants import PF_VERSION_CHECK
+from promptflow._constants import PF_VERSION_CHECK, LAST_CHECK_TIME
 from promptflow._sdk._constants import HOME_PROMPT_FLOW_DIR, PROMPT_FLOW_HOME_DIR_ENV_VAR
 from promptflow._sdk._errors import GenerateFlowToolsJsonError
 from promptflow._sdk._telemetry.logging_handler import get_scrubbed_cloud_role
@@ -38,7 +40,8 @@ from promptflow._sdk._utils import (
     snake_to_camel,
 )
 from promptflow._utils.load_data import load_data
-from promptflow._utils.version_hint_utils import hint_for_update, check_latest_version
+from promptflow._utils.version_hint_utils import check_latest_version
+
 
 TEST_ROOT = Path(__file__).parent.parent.parent
 CONNECTION_ROOT = TEST_ROOT / "test_configs/connections"
@@ -202,18 +205,30 @@ class TestUtils:
         for thread in threads:
             thread.join()
 
-    @pytest.mark.parametrize("concurrent_count", [1, 2, 4, 8])
-    def test_concurrent_hint_for_update(self, concurrent_count):
-        from concurrent.futures import ThreadPoolExecutor
-        with patch('promptflow._utils.version_hint_utils.datetime') as mock_datetime:
-            mock_datetime.datetime.now.return_value = datetime.datetime.now()
-            mock_datetime.datetime.strptime.return_value = datetime.datetime.now() - datetime.timedelta(days=8)
-            mock_datetime.timedelta.return_value = datetime.timedelta(days=7)
-            hint_for_update()
-            with ThreadPoolExecutor(max_workers=concurrent_count) as pool:
-                pool.submit(check_latest_version)
+    def test_concurrent_hint_for_update(self):
+        def mock_check_latest_version():
+            time.sleep(5)
+            check_latest_version()
 
+        with patch('promptflow._utils.version_hint_utils.datetime') as mock_datetime, patch(
+                "promptflow._utils.version_hint_utils.check_latest_version", side_effect=mock_check_latest_version):
+            from promptflow._sdk._telemetry import monitor_operation
+
+            class HintForUpdate:
+                @monitor_operation(activity_name="pf.flows.test")
+                def hint_func(self):
+                    return
+            current_time = datetime.datetime.now()
+            mock_datetime.datetime.now.return_value = current_time
+            mock_datetime.datetime.strptime.return_value = current_time - datetime.timedelta(days=8)
+            mock_datetime.timedelta.return_value = datetime.timedelta(days=7)
+            HintForUpdate().hint_func()
             assert Path(HOME_PROMPT_FLOW_DIR / PF_VERSION_CHECK).exists()
+            with open(HOME_PROMPT_FLOW_DIR / PF_VERSION_CHECK, "r") as f:
+                cached_versions = json.load(f)
+            # since mock_check_latest_version is a demon thread, it will exit when main thread complete, so
+            # LAST_CHECK_TIME won't be updated since sleep 5s
+            assert LAST_CHECK_TIME not in cached_versions or cached_versions[LAST_CHECK_TIME] != str(current_time)
 
     @pytest.mark.parametrize(
         "data_path",
@@ -256,8 +271,10 @@ class TestUtils:
         df = load_data(data_path)
         assert len(df) == 10000
         # specify max_rows_count
-        df = load_data(data_path, max_rows_count=5000)
-        assert len(df) == 5000
+        max_rows_count = 5000
+        head_rows = load_data(data_path, max_rows_count=max_rows_count)
+        assert len(head_rows) == max_rows_count
+        assert head_rows == df[:max_rows_count]
 
     @pytest.mark.parametrize(
         "script_name, expected_result",